refactor(logger): reorganize logger service into dedicated module
Move logger service implementation and interfaces into a dedicated logger module
Update all imports to reflect new module structure
Add comprehensive tests for logger service functionality
